api/requirements.txt
google-cloud-storage
google-cloud-secret-manager

falcon
requests

flake8
black
pylint